Python/requirements.txt
ijson
openai
sounddevice
soundfile
keyboard
pydub
pyserial
numpy